refactor: 쓸데없는 계층(Storage) 삭제 및 그로인한 전면 수정
