+++ ProjectManager/ProjectManager/Domain/Repository/PersistentRepository.swift
@@ -9,7 +9,6 @@ import RxSwift
 import RxRelay
 
 protocol PersistentRepositoryProtocol {
-    var persistentStorage: PersistentStorageProtocol { get }
     func create(projectEntity: ProjectEntity)
     func create(projectEntities: [ProjectEntity])
     func read() -> BehaviorRelay<[ProjectEntity]>
@@ -19,40 +18,155 @@ protocol PersistentRepositoryProtocol {
     func deleteAll()
 }
 
-struct PersistentRepository: PersistentRepositoryProtocol {
-    let persistentStorage: PersistentStorageProtocol
+extension PersistentRepositoryProtocol {
+    func parse(from project: Project) -> ProjectEntity? {
+        guard let id = project.id,
+              let status = ProjectStatus.convert(statusString: project.status),
+              let title = project.title,
+              let deadline = project.deadline,
+              let body = project.body else {
+            return nil
+        }
+        
+        return ProjectEntity(
+            id: id,
+            status: status,
+            title: title,
+            deadline: deadline,
+            body: body
+        )
+    }
     
-    init(persistentStorage: PersistentStorageProtocol) {
-        self.persistentStorage = persistentStorage
+    func parse(from projectContent: ProjectEntity) -> ProjectDTO? {
+        return ProjectDTO(
+            id: projectContent.id.uuidString,
+            status: projectContent.status.string,
+            title: projectContent.title,
+            deadline: projectContent.deadline,
+            body: projectContent.body
+        )
+    }
+}
+
+final class PersistentRepository: PersistentRepositoryProtocol {
+    private let persistentManager: PersistentManagerProtocol
+    private lazy var projectEntities = BehaviorRelay<[ProjectEntity]>(value: fetchCoreDate())
+    
+    init(persistentManager: PersistentManagerProtocol) {
+        self.persistentManager = persistentManager
+    }
+    
+    private func fetchCoreDate() -> [ProjectEntity] {
+        let currentProjects = persistentManager.read()
+        let contents = currentProjects.compactMap { parse(from: $0) }
+        
+        return contents
     }
 }
 
 extension PersistentRepository {
     func create(projectEntity: ProjectEntity) {
-        persistentStorage.create(projectEntity: projectEntity)
+        createCoreDate(newProjectEntity: projectEntity)
+        createProjectEntities(newProjectEntity: projectEntity)
     }
     
     func create(projectEntities: [ProjectEntity]) {
-        persistentStorage.create(projectEntities: projectEntities)
+        createCoreDate(newProjectEntities: projectEntities)
+        createProjectEntities(newProjectEntities: projectEntities)
     }
     
     func read() -> BehaviorRelay<[ProjectEntity]> {
-        return persistentStorage.read()
+        return projectEntities
     }
     
     func read(projectEntityID: UUID?) -> ProjectEntity? {
-        return persistentStorage.read(id: projectEntityID)
+        return projectEntities.value.filter { $0.id == projectEntityID }.first
     }
     
     func update(projectEntity: ProjectEntity) {
-        persistentStorage.update(projectEntity: projectEntity)
+        updateCoreDate(newProjectEntity: projectEntity)
+        updateProjectEntities(newProjectEntity: projectEntity)
     }
     
     func delete(projectEntityID: UUID?) {
-        persistentStorage.delete(projectEntityID: projectEntityID)
+        deleteCoreDate(projectEntityID: projectEntityID)
+        deleteProjectEntities(projectEntityID: projectEntityID)
     }
     
     func deleteAll() {
-        persistentStorage.deleteAll()
+        deleteAllCoreDate()
+        deleteAllProjectEntities()
     }
 }
+
+extension PersistentRepository {
+    private func createCoreDate(newProjectEntity: ProjectEntity) {
+        guard let newProject = parse(from: newProjectEntity) else {
+            return
+        }
+        
+        persistentManager.create(project: newProject)
+    }
+    
+    private func createProjectEntities(newProjectEntity: ProjectEntity) {
+        var projects = projectEntities.value
+        
+        projects.append(newProjectEntity)
+        projectEntities.accept(projects)
+    }
+    
+    private func createCoreDate(newProjectEntities: [ProjectEntity]) {
+        let newProjects = newProjectEntities.compactMap {
+            parse(from: $0)
+        }
+        
+        persistentManager.create(projects: newProjects)
+    }
+    
+    private func createProjectEntities(newProjectEntities: [ProjectEntity]) {
+        projectEntities.accept(newProjectEntities)
+    }
+    
+    private func updateCoreDate(newProjectEntity: ProjectEntity) {
+        guard let newProject = parse(from: newProjectEntity) else {
+            return
+        }
+        
+        persistentManager.update(project: newProject)
+    }
+    
+    private func updateProjectEntities(newProjectEntity: ProjectEntity) {
+        let projects = projectEntities.value
+        
+        if let indexToUpdated = projects.firstIndex(where: { $0.id == newProjectEntity.id}) {
+            var projectsToUpdate = projectEntities.value
+            
+            projectsToUpdate[indexToUpdated] = newProjectEntity
+            projectEntities.accept(projectsToUpdate)
+        }
+    }
+    
+    private func deleteCoreDate(projectEntityID: UUID?) {
+        persistentManager.delete(projectEntityID: projectEntityID)
+    }
+    
+    private func deleteProjectEntities(projectEntityID: UUID?) {
+        let projects = projectEntities.value
+        
+        if let indexToDelete = projects.firstIndex(where: { $0.id == projectEntityID}) {
+            var projectsToDelete = projectEntities.value
+            
+            projectsToDelete.remove(at: indexToDelete)
+            projectEntities.accept(projectsToDelete)
+        }
+    }
+    
+    private func deleteAllCoreDate() {
+        persistentManager.deleteAll()
+    }
+    
+    private func deleteAllProjectEntities() {
+        projectEntities.accept([])
+    }
+}
+
