+++ ProjectManager/ProjectManagerTests/Data/Mock/MockNetworkManager.swift
@@ -11,7 +11,13 @@ import RxSwift
 struct SampleData {
     static private let deadline = Calendar.current.date(byAdding: .day, value: 1, to: Date())!
 
-    static let data = ProjectDTO(id: UUID().uuidString, status: ProjectStatus.todo.string, title: "title111", deadline: String(deadline.timeIntervalSince1970), body: "bodyttt")
+    static let data = ProjectDTO(
+        id: UUID().uuidString,
+        status: ProjectStatus.todo.string,
+        title: "title111",
+        deadline: String(deadline.timeIntervalSince1970),
+        body: "bodyttt"
+    )
 }
 
 final class MockFirebase {
@@ -38,12 +44,29 @@ final class MockFirebase {
     }
 }
 
-final class MockNetworkManager {
+final class MockNetworkRepository {
     let database = MockFirebase(error: nil)
 }
 
-extension MockNetworkManager: NetworkManagerProtocol {
-    func read() -> Observable<[ProjectDTO]> {
+extension MockNetworkRepository: NetworkRepositoryProtocol {
+    func update(repository: PersistentRepositoryProtocol) {
+        let projects = repository.read().value.compactMap {
+            parse(from: $0)
+        }
+        
+        update(projects: projects)
+    }
+    
+    func read(repository: PersistentRepositoryProtocol) -> Disposable {
+        return read()
+            .subscribe(onNext: {[weak self] data in
+                self?.synchronize(with: data, to: repository)
+            })
+    }
+}
+
+extension MockNetworkRepository {
+    private func read() -> Observable<[ProjectDTO]> {
         
         return Observable.create { [weak self] emitter in
             self?.database.getData { _, snapshot in
@@ -60,13 +83,31 @@ extension MockNetworkManager: NetworkManagerProtocol {
         }
     }
     
-    func update(projects: [ProjectDTO]) {
+    private func update(projects: [ProjectDTO]) {
         var newData: [String: [String: String]] = [:]
         
         projects.forEach {
-            newData[$0.id] = ["id": $0.id, "status": $0.status, "title": $0.title, "deadline": $0.deadline, "body": $0.body]
+            newData[$0.id] = [
+                "id": $0.id,
+                "status": $0.status,
+                "title": $0.title,
+                "deadline": $0.deadline,
+                "body": $0.body
+            ]
         }
         
         database.setValue(newData)
+    }
+    
+    private func synchronize(
+        with projects: [ProjectDTO],
+        to repository: PersistentRepositoryProtocol
+    ) {
+        let formattedProjects = projects.compactMap {
+            parse(from: $0)
+        }
+        
+        repository.deleteAll()
+        repository.create(projectEntities: formattedProjects)
     }
 }
